[Edit]Refactering challenge

+++ Desktop/JavaMasterClass/ReadingUserInput/src/ReadingUserInputChallenge.java
@@ -4,13 +4,14 @@ public class ReadingUserInputChallenge {
         Scanner scanner = new Scanner(System.in);
 
         int counter = 1;
-        int sum = 0;
+        double sum = 0;
 
         while(counter <= 5){
             System.out.println("Enter number #" + counter + ":");
             String nextNumber = scanner.nextLine();
             try {
-                int number = Integer.parseInt(nextNumber);
+//                int number = Integer.parseInt(nextNumber);
+                double number = Double.parseDouble(nextNumber);
                 counter++;
                 sum += number;
             } catch (NumberFormatException nfe){
